inputエリアをcomponent分割
