+++ app/components/InputTodo.module.scss
@@ -1,0 +1,53 @@
+@use "../variables.module.scss" as vars;
+
+.form {
+  padding: 8px 16px;
+  margin-bottom: 32px;
+  border: 1px solid vars.$borderColor;
+  border-radius: 8px;
+  display: block;
+  margin-top: 0em;
+  unicode-bidi: isolate;
+  .formGroups {
+    display: flex;
+    .inputTask {
+      flex: 4 1;
+      padding: 8px 16px;
+      .inputLabel {
+        display: block;
+        font-size: 12px;
+        margin-bottom: 4px;
+      }
+      .input {
+        padding: 8px;
+        border-radius: 8px;
+        border: 1px solid vars.$borderColor;
+        width: 100%;
+      }
+    }
+    .inputDate {
+      flex: 1 1;
+      padding: 8px 16px;
+      .inputLabel {
+        display: block;
+        font-size: 12px;
+        margin-bottom: 4px;
+      }
+      .input {
+        padding: 8px;
+        border-radius: 8px;
+        border: 1px solid vars.$borderColor;
+        width: 100%;
+      }
+    }
+  }
+  .formFooter {
+    display: flex;
+    justify-content: flex-end;
+    padding: 10px;
+    .button {
+      background-color: vars.$primartColor;
+      color: vars.$fontColorLight;
+    }
+  }
+}
